+++ src/com.cc
@@ -29,7 +29,7 @@ set_appid (IShellLink *sl, lisp lappid)
   char *b = (char *)alloca (xstring_length (lappid) * 2 + 1);
   w2s (b, lappid);
 
-  int l = (strlen (b) + 1);
+  long long l = (strlen (b) + 1);
   wchar_t *w = (wchar_t *)alloca (l * sizeof (wchar_t));
   MultiByteToWideChar (CP_ACP, 0, b, -1, w, l);
 
@@ -89,7 +89,7 @@ Fcreate_shortcut (lisp lobject, lisp llink, lisp keys)
       ole_error (SHGetDesktopFolder (&sf));
 
       map_sl_to_backsl (path);
-      int l = (strlen (path) + 1);
+      long long l = (strlen (path) + 1);
       wchar_t *w = (wchar_t *)alloca (l * sizeof (wchar_t));
       MultiByteToWideChar (CP_ACP, 0, path, -1, w, l);
 
@@ -119,7 +119,7 @@ Fcreate_shortcut (lisp lobject, lisp llink, lisp keys)
   ole_error (sl->QueryInterface (IID_IPersistFile, (void **)&pf));
 
   pathname2cstr (llink, path);
-  int l = (strlen (path) + 1);
+  long long l = (strlen (path) + 1);
   wchar_t *w = (wchar_t *)alloca (l * sizeof (wchar_t));
   MultiByteToWideChar (CP_ACP, 0, path, -1, w, l);
   ole_error (pf->Save (w, 1));
@@ -192,7 +192,7 @@ Fresolve_shortcut (lisp lshortcut)
   char shortcut[PATH_MAX + 1];
   pathname2cstr (lshortcut, shortcut);
   map_sl_to_backsl (shortcut);
-  int l = (strlen (shortcut) + 1);
+  long long l = (strlen (shortcut) + 1);
   wchar_t *w = (wchar_t *)alloca (l * sizeof (wchar_t));
   MultiByteToWideChar (CP_ACP, 0, shortcut, -1, w, l);
 
@@ -239,7 +239,7 @@ Fole_drop_files (lisp lpath, lisp lclsid, lisp ldir, lisp lfiles)
   char dir[PATH_MAX + 1];
   pathname2cstr (ldir, dir);
   map_sl_to_backsl (dir);
-  int maxl = strlen (dir);
+  long long maxl = strlen (dir);
 
   lisp f = lfiles;
   int nfiles;
